DowntimeForm.js almost completed
DowntimeForm.js can generate new entries into the array. You must refresh the page to see the changes, although after enough refreshes, the data can be lost. there is also a bug that causes the Site and Telescope entries when 'North' and 'A' are selected by default.
Also, the use of Localstorage is being used from now on.